Legacy Parser Cleanup (#59)
For a little while now, Molt has had two TCL parsers, one that parses to an internal form for later evaluation, and one that parses and evaluates in one step. The latter was only used by the expression parser for interpolated TCL code. This pull request moves the remains of the legacy parser to expr.rs, and revises it to parse using the new parser before evaluating.