Added parser::parse_varname_literal
parser::parse_varname_literal parses a string that is known to be a variable name.  The only question is whether it has array syntax or not.  If it does, it's an array reference; if not, it's a scalar.  No substitutions are done.  The result is a VarName; no errors are possible.

Updated `Value::as_var_name` to use it.