Change Datum into an enum
Using an enum instead of a struct makes it impossible to use the wrong
field, and allows us to couple the tag-based dispatch with uses of
those fields.  This pays off significantly in type coercions for each
operator, where some new helper functions have been used to unify the
coercion and evaluation blocks.